feat(): updated examples, moved to separate directories
