+++ examples/REST/rest-usdm-order.ts
@@ -1,4 +1,4 @@
-import { USDMClient } from '../src/index';
+import { USDMClient } from '../../src/index';
 
 // or
 // import { USDMClient } from 'binance';
@@ -10,13 +10,6 @@ const client = new USDMClient({
   api_secret: secret,
   api_key: key,
   beautifyResponses: true,
-  /**
-   * Note: testnet is NOT a good place to test strategy performance.
-   *
-   * For more information and guidance, refer to:
-   * https://github.com/tiagosiebler/awesome-crypto-examples/wiki/CEX-Testnets
-   */
-  testnet: true,
 });
 
 async function start() {
