+++ examples/REST/rest-usdm-public.ts
@@ -1,4 +1,4 @@
-import { USDMClient } from '../src';
+import { USDMClient } from '../../src';
 
 const client = new USDMClient({
   testnet: true,
