+++ examples/WEBSOCKET/ws-proxy-socks.ts
@@ -1,7 +1,7 @@
 /**
  * Minimal example for using a socks proxy with the ws client, extracted from https://github.com/tiagosiebler/binance/pull/319
  */
-import { WebsocketClient } from '../src';
+import { WebsocketClient } from '../../src/index';
 
 // or
 // import { WebsocketClient } from 'binance';
