+++ examples/util/auth/rest-private-ed25519.md
